+++ ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: paminna <paminna@student.21-school.ru>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/01 21:25:46 by paminna           #+#    #+#             */
-/*   Updated: 2020/11/01 21:40:07 by paminna          ###   ########.fr       */
+/*   Updated: 2020/11/03 19:58:37 by paminna          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@
 
 int ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while(i < n)
